requirements.txt
numpy==1.18.5
scipy==1.1.0
Pillow==7.2.0
cython==0.29.21
matplotlib==3.3.2
scikit-image==0.17.2
tensorflow-gpu==1.15.0
keras==2.0.8
opencv-python==4.4.0.42
h5py==2.10.0
imgaug==0.4.0
IPython[all]